+++ NEWS.md
@@ -5,6 +5,23 @@ New features:
  - Render torque tiles server side (WIP)
 
 
+# Version 0.32.4
+2014-12-15
+
+Bugfixes:
+ - Torque Infinity number of steps (#233)
+
+
+# Version 0.32.3
+2014-12-15
+
+Bugfixes:
+ - Fix incorrect number of steps in Torque (see https://github.com/CartoDB/torque/issues/60)
+
+Enhancements:
+ - Split torque renderer functionality into several files
+
+
 # Version 0.32.2
 2014-12-11